# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Benchwarmer management API
  version: "1"
  summary: Management API for the Benchwarmer opportunistic GPU worker.
  description: |
    The management listener (`listen.management`, loopback by default) serves
    `/api/v1/*` and `/metrics`. The inference listener (`listen.inference`)
    serves the OpenAI-compatible `/v1/*` proxy; its availability contract is
    documented under the `/v1/{path}` operation below.

    Times are RFC 3339 with an offset. Clients render them in local time.
    Fields whose names end in `_ns` are nanoseconds; `_ms` are milliseconds.

    ## Authentication (ADR 0007)

    Bearer tokens in `Authorization: Bearer <token>`. There are three
    independent tokens, each stored in its own file under the data
    directory's `secrets` folder:

    | Token | Accepted on |
    |---|---|
    | management | every management endpoint |
    | agent | `GET /api/v1/status`, `GET /api/v1/health`, `PUT /api/v1/mode`, `POST /api/v1/agent/report` |
    | inference | the inference proxy only; refused (403) on the management API |

    The agent token is readable by interactive users so the tray and session
    agent work for a non-administrator; it cannot read or change config,
    applications, or events, drain, reload, or read metrics.

    Tokenless access: when `security.loopback_trust` is true, `GET`
    endpoints (including `/metrics`) are allowed without a token if the TCP
...
    Every write needs a token regardless of source.

...

...

...
servers:
...
security:
  - bearer: []
  - {}
...
paths:
...
  /api/v1/status:
    get:
      tags: [status]
      operationId: getStatus
      summary: Current condition, state, decision, telemetry, and timers.
      description: Loopback trust, management token, or agent token.
      responses:
        "200":
          description: Status snapshot.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Status" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/config:
    get:
      tags: [config]
      operationId: getConfig
      summary: Active configuration with secrets redacted.
      description: |
        Loopback trust or management token. Secret-looking values in
        `runtime.args` (for example the value after `--api-key`) are
        replaced by `"<redacted>"`.
      responses:
        "200":
          description: Redacted config and where it was loaded from.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConfigResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    put:
      tags: [config]
      operationId: putConfig
      summary: Replace the configuration.
      description: |
        Management token only. The body is a complete `Config`. Values
        returned as `"<redacted>"` by `GET` may be sent back unchanged and
        keep their stored secret, so a client can round-trip what it read.
        The config is validated, persisted atomically, applied, and audited.
        The response reports whether the change needs a runtime reload or a
        service restart.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Config" }
      responses:
        "200":
          description: Applied. Config is redacted; `impact` is present.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConfigResponse" }
        "400": { $ref: "#/components/responses/InvalidConfig" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Rejected" }
        "413": { $ref: "#/components/responses/TooLarge" }
  /api/v1/events:
    get:
      tags: [status]
      operationId: getEvents
      summary: Recent audit events, newest first.
      description: Loopback trust or management token.
      parameters:
        - name: limit
          in: query
          description: Page size. Default 100; values above 1000 are treated as 1000.
          schema: { type: integer, minimum: 1, default: 100 }
        - name: before_id
          in: query
          description: Return only events with a smaller id (use `next_before_id` from the previous page).
          schema: { type: integer, format: int64, minimum: 1 }
        - name: type
          in: query
          description: Event type filter. Repeat the parameter or separate with commas; at most 64 types.
          style: form
          explode: true
          schema:
            type: array
            items: { $ref: "#/components/schemas/EventType" }
      responses:
        "200":
          description: A page of events.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EventsResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/applications:
    get:
      tags: [config]
      operationId: getApplications
      summary: Application classification rules, in evaluation order.
      description: Loopback trust or management token.
      responses:
        "200":
          description: The rules.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationsBody" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    put:
      tags: [config]
      operationId: putApplications
      summary: Replace the application rules.
      description: |
        Management token only. Replaces `applications` in the active config
        and applies it like `PUT /api/v1/config`. `applications` is required;
        send `[]` for none.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApplicationsBody" }
      responses:
        "200":
          description: Applied rules.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationsBody" }
        "400": { $ref: "#/components/responses/InvalidConfig" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Rejected" }
        "413": { $ref: "#/components/responses/TooLarge" }
  /api/v1/mode:
    put:
      tags: [control]
      operationId: putMode
      summary: Set the manual mode (Auto, Pause AI, AI Priority).
      description: |
        Management token or agent token. `duration` is a Go duration string
        (`"30m"`, `"2h"`), empty for indefinite, `"until_reboot"`, or
        `"until_tomorrow"`. `auto` takes no duration. Pause durations are at
        most 720h; AI Priority durations at most `modes.max_ai_priority`.
        `set_by` defaults to `"api"`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ModeRequest" }
      responses:
        "200":
          description: The mode now in force.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ModeStatus" }
        "400":
          description: Invalid mode or duration (`invalid_mode`) or malformed body (`bad_request`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Rejected" }
        "413": { $ref: "#/components/responses/TooLarge" }
  /api/v1/drain:
    post:
      tags: [control]
      operationId: postDrain
      summary: Drain and unload now (manual; not counted toward anti-thrashing).
      description: Management token only. The worker reloads when policy allows.
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ActionRequest" }
      responses:
        "202":
          description: Accepted; the controller acts asynchronously.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ActionResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Rejected" }
  /api/v1/reload:
    post:
      tags: [control]
      operationId: postReload
      summary: Restart the runtime (drains first).
      description: Management token only.
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ActionRequest" }
      responses:
        "202":
          description: Accepted; the controller acts asynchronously.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ActionResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Rejected" }
  /api/v1/agent/report:
    post:
      tags: [agent]
      operationId: postAgentReport
      summary: Session agent report (foreground app, fullscreen, idle, lock).
      description: Agent token or management token. Unknown fields are ignored.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AgentReport" }
      responses:
        "204":
          description: Recorded.
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "413": { $ref: "#/components/responses/TooLarge" }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: Management, agent, or inference token (see the access table above).
...
  responses:
    BadRequest:
      description: Malformed request (`bad_request`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    InvalidConfig:
      description: |
        `invalid_config` with one `details` entry per problem, or
        `bad_request` for malformed JSON or unknown fields.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: configuration is invalid
            code: invalid_config
            details:
              - field: runtime.host
                message: 'must be a loopback address; the runtime must not be reachable from other machines (got "0.0.0.0")'
    Unauthorized:
      description: No token, or the token is not valid (`unauthorized`, `invalid_token`).
      headers:
        WWW-Authenticate:
          schema: { type: string, const: 'Bearer realm="benchwarmer"' }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: A valid token that this endpoint does not accept (`forbidden`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Rejected:
      description: The service refused the request in its current state (`rejected`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    TooLarge:
      description: Request body over the limit (`body_too_large`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    State:
      type: string
      enum: [STOPPED, COOLDOWN, LOADING, READY, BUSY, DRAINING, PREEMPTING, SUPPRESSED, DISABLED, ERROR]
    Condition:
      type: string
      enum: [Available, Yielding, Unavailable]
...
    EventType:
      type: string
      enum:
        - service_started
        - service_stopping
        - state_changed
        - decision_changed
        - load_started
        - load_completed
        - load_failed
        - drain_started
        - drain_completed
        - grace_shortened
        - preempted
        - runtime_stopped
        - runtime_crashed
        - kill_failed
        - vram_released
        - vram_release_timeout
        - orphan_killed
        - cooldown_started
        - cooldown_reset
        - suppression_started
        - recovery_scheduled
        - mode_changed
        - mode_expired
        - config_changed
        - config_rejected
        - config_recovered
        - telemetry_lost
        - telemetry_recovered
        - telemetry_degraded
        - power_suspend
        - power_resume
        - session_change
        - agent_connected
        - agent_disconnected
        - device_lost
        - requests_rejected
        - request_force_closed
...
    Error:
      type: object
      description: Every non-2xx JSON body, including the inference proxy's 503.
      required: [error, code]
      properties:
        error: { type: string, description: Human-readable message. }
        code: { $ref: "#/components/schemas/ErrorCode" }
        details:
          type: array
          items: { $ref: "#/components/schemas/FieldError" }
        condition: { $ref: "#/components/schemas/Condition" }
        reason: { type: string }
        retry_after: { type: string, format: date-time }
...
    Status:
      type: object
      required: [time, version, condition, state, summary, mode, profile, runtime, gpu, decision, timers, agent, config_source]
      properties:
        time: { type: string, format: date-time }
        version: { type: string }
        condition: { $ref: "#/components/schemas/Condition" }
        state: { $ref: "#/components/schemas/State" }
        summary: { type: string, description: One sentence for the tray tooltip and dashboard header. }
        mode: { $ref: "#/components/schemas/ModeStatus" }
        profile: { $ref: "#/components/schemas/ProfileStatus" }
        runtime: { $ref: "#/components/schemas/RuntimeStatus" }
        gpu: { $ref: "#/components/schemas/GPUStatus" }
        decision: { $ref: "#/components/schemas/Decision" }
        timers: { $ref: "#/components/schemas/TimerStatus" }
        agent: { $ref: "#/components/schemas/AgentStatus" }
        trigger: { type: string, description: "Competing application or metric, when known." }
        recent_errors:
          type: array
          items: { $ref: "#/components/schemas/Event" }
        config_source:
          type: string
          enum: [primary, last_good, default]
          description: Anything but primary is a problem the UI must show.
    ModeStatus:
      type: object
      required: [mode]
      properties:
        mode: { $ref: "#/components/schemas/Mode" }
        until: { type: string, format: date-time }
        set_by: { type: string }
        set_at: { type: string, format: date-time }
    ProfileStatus:
      type: object
      required: [active, source, timezone]
      properties:
        active: { type: string }
        source: { type: string, description: '"default", "schedule:<name>", or "ai_priority".' }
        timezone: { type: string }
        next_change: { type: string, format: date-time }
    RuntimeStatus:
      type: object
      required: [ready, model, active_requests, oldest_request_seconds, crash_count]
      properties:
        pid: { type: integer }
        ready: { type: boolean }
        model: { type: string, description: Model file name only. }
        active_requests: { type: integer }
        oldest_request_seconds: { type: number }
        loaded_at: { type: string, format: date-time }
        last_load_seconds: { type: number }
        footprint_mib: { type: integer }
        crash_count: { type: integer }
    GPUStatus:
      type: object
      required: [adapter, confidence, total_util_pct, own_util_pct, external_util_pct, vram_total_mib, vram_used_mib, vram_free_mib, own_vram_mib, external_vram_mib]
      properties:
        adapter: { type: string }
        confidence: { $ref: "#/components/schemas/Confidence" }
        total_util_pct: { type: number }
        own_util_pct: { type: number }
        external_util_pct: { type: number }
        vram_total_mib: { type: integer }
        vram_used_mib: { type: integer }
        vram_free_mib: { type: integer }
        own_vram_mib: { type: integer }
        external_vram_mib: { type: integer }
        temperature_c: { type: number, description: Absent when unknown. }
        top_external:
          type: array
          items: { $ref: "#/components/schemas/AppMatch" }
        sampled_at: { type: string, format: date-time }
...
    AgentStatus:
      type: object
      required: [connected]
      properties:
        connected: { type: boolean }
        last_report: { type: string, format: date-time }
        session_id: { type: integer, format: uint32 }
...
    EventsResponse:
      type: object
      required: [events]
      properties:
        events:
          type: array
          items: { $ref: "#/components/schemas/Event" }
        next_before_id:
          type: integer
          format: int64
          description: Pass as before_id for the next (older) page; absent when there are no older events.
    ModeRequest:
      type: object
      required: [mode]
      additionalProperties: false
      properties:
        mode: { $ref: "#/components/schemas/Mode" }
        duration:
          type: string
          description: Go duration, empty for indefinite, "until_reboot", or "until_tomorrow".
          examples: ["30m", "2h", "until_reboot", "until_tomorrow"]
        set_by: { type: string, maxLength: 64, examples: ["tray", "ui"] }
    ActionRequest:
      type: object
      additionalProperties: false
      properties:
        reason: { type: string, maxLength: 200 }
    ActionResponse:
      type: object
      required: [accepted, action, reason]
      properties:
        accepted: { type: boolean }
        action: { type: string, enum: [drain, reload] }
        reason: { type: string }
    AgentReport:
      type: object
      required: [session_id, fullscreen, idle_seconds, locked]
      properties:
        session_id: { type: integer, format: uint32 }
        foreground_pid: { type: integer, format: uint32 }
        foreground_name: { type: string }
        foreground_path: { type: string }
        fullscreen: { type: boolean }
        notification_state: { type: string }
        idle_seconds: { type: number }
        locked: { type: boolean }
    ConfigResponse:
      type: object
      required: [config, source]
      properties:
        config: { $ref: "#/components/schemas/Config" }
        source: { type: string, enum: [primary, last_good, default] }
        impact: { $ref: "#/components/schemas/Impact" }
...
    ApplicationsBody:
      type: object
      required: [applications]
      additionalProperties: false
      properties:
        applications:
          type: array
          items: { $ref: "#/components/schemas/AppRule" }

    # Configuration (internal/config). Unknown fields are rejected.
    Config:
      type: object
      additionalProperties: false
      required: [schema_version, runtime, listen, telemetry, safety, contention, profiles, default_profile, ai_priority_profile, schedules, timezone, applications, anti_thrash, recovery, modes, signals, security, retention, metrics, logging]
      properties:
        schema_version: { type: integer, const: 1 }
        runtime: { $ref: "#/components/schemas/RuntimeConfig" }
        listen: { $ref: "#/components/schemas/ListenConfig" }
        telemetry: { $ref: "#/components/schemas/TelemetryConfig" }
        safety: { $ref: "#/components/schemas/SafetyConfig" }
        contention: { $ref: "#/components/schemas/ContentionConfig" }
        profiles:
          type: object
          additionalProperties: { $ref: "#/components/schemas/Profile" }
        default_profile: { type: string }
        ai_priority_profile: { type: string }
        schedules:
          type: array
          items: { $ref: "#/components/schemas/Schedule" }
        timezone: { type: string, examples: ["America/Chicago"] }
        applications:
          type: array
          items: { $ref: "#/components/schemas/AppRule" }
        anti_thrash: { $ref: "#/components/schemas/AntiThrashConfig" }
        recovery: { $ref: "#/components/schemas/RecoveryConfig" }
        modes: { $ref: "#/components/schemas/ModesConfig" }
        signals: { $ref: "#/components/schemas/SignalsConfig" }
        security: { $ref: "#/components/schemas/SecurityConfig" }
        retention: { $ref: "#/components/schemas/RetentionConfig" }
        metrics: { $ref: "#/components/schemas/MetricsConfig" }
        logging: { $ref: "#/components/schemas/LoggingConfig" }
...